Split trigger dispatcher modules
